+++ lib/image_util/generator.rb
@@ -3,5 +3,6 @@
 module ImageUtil
   module Generator
     autoload :BitmapText, "image_util/generator/bitmap_text"
+    autoload :Example, "image_util/generator/example"
   end
 end